requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
beautifulsoup4
openai
